trait mearnings - refactoring modules
