+++ tutorials/src/lib.rs
@@ -8,9 +8,7 @@ fn another_private_function() {
 }
 
 pub mod front_of_house;
-pub mod news_article;
-pub mod summary;
-pub mod tweet;
+pub mod media_aggrigator;
 
 pub fn eat_at_restaurant() {
     crate::front_of_house::hosting::add_to_waitlist();
