+++ tutorials/src/main.rs
@@ -1,5 +1,4 @@
-use tutorials::news_article::run_news_article;
-use tutorials::tweet::run_tweets;
+use tutorials::media_aggrigator::{news_article::run_news_article, tweet::run_tweets};
 
 fn main() {
     run_tweets();
