+++ tutorials/src/media_aggrigator/news_article.rs
@@ -1,4 +1,4 @@
-use crate::summary::Summary;
+use super::summary::Summary;
 
 pub struct NewsArticle {
     pub headline: String,
